Task 1

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("Size of char: %lu byte(s)\n", sizeof(c));
-	printf("Size of int: %lu byte(s)\n", sizeof(i));
-	printf("Size of long int: %lu byte(s)\n", sizeof(li));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(lli));
-	printf("Size of float: %lu byte(s)\n", sizeof(f));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(li));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(lli));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
